+++ src/op/iter_all_edges.rs
@@ -66,6 +66,14 @@ impl IterAllEdges for Vec<(usize, usize)> {
     }
 }
 
+// Slice
+
+impl IterAllEdges for [(usize, usize)] {
+    fn iter_all_edges(&self) -> impl Iterator<Item = &(usize, usize)> {
+        self.iter()
+    }
+}
+
 // Arr
 
 impl<const V: usize> IterAllEdges for [(usize, usize); V] {
@@ -95,6 +103,17 @@ mod tests {
     #[test]
     fn vec() {
         let graph = vec![(0, 1), (1, 2), (2, 0)];
+        let mut iter = graph.iter_all_edges();
+
+        assert_eq!(iter.next(), Some(&(0, 1)));
+        assert_eq!(iter.next(), Some(&(1, 2)));
+        assert_eq!(iter.next(), Some(&(2, 0)));
+        assert_eq!(iter.next(), None);
+    }
+
+    #[test]
+    fn slice() {
+        let graph: &[(usize, usize)] = &[(0, 1), (1, 2), (2, 0)];
         let mut iter = graph.iter_all_edges();
 
         assert_eq!(iter.next(), Some(&(0, 1)));