+++ src/op/iter_all_weighted_edges.rs
@@ -74,6 +74,17 @@ impl<W> IterAllWeightedEdges<W> for Vec<(usize, usize, W)> {
     }
 }
 
+// Slice
+
+impl<W> IterAllWeightedEdges<W> for [(usize, usize, W)] {
+    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = &(usize, usize, W)>
+    where
+        W: 'a,
+    {
+        self.iter()
+    }
+}
+
 // Arr
 
 impl<const V: usize, W> IterAllWeightedEdges<W> for [(usize, usize, W); V] {
@@ -109,6 +120,17 @@ mod tests {
     #[test]
     fn vec() {
         let graph = vec![(0, 1, 2), (1, 2, 3), (2, 0, 4)];
+        let mut iter = graph.iter_all_weighted_edges();
+
+        assert_eq!(iter.next(), Some(&(0, 1, 2)));
+        assert_eq!(iter.next(), Some(&(1, 2, 3)));
+        assert_eq!(iter.next(), Some(&(2, 0, 4)));
+        assert_eq!(iter.next(), None);
+    }
+
+    #[test]
+    fn slice() {
+        let graph: &[(usize, usize, usize)] = &[(0, 1, 2), (1, 2, 3), (2, 0, 4)];
         let mut iter = graph.iter_all_weighted_edges();
 
         assert_eq!(iter.next(), Some(&(0, 1, 2)));